test(feedback): fix FeedbackServiceTest namespace after User→Feedback split
The test moved to tests/Unit/Plugins/Feedback/ but still declared the
Tests\Unit\Plugins\User namespace and referenced a FakeFeedbackStore
under User\Support that no longer exists, causing 11 CI errors. Point it
at Tests\Unit\Plugins\Feedback\Support\FakeFeedbackStore.
