+++ tests/Unit/Plugins/Feedback/Support/FakeFeedbackStore.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\Plugins\User\Support;
+namespace Tests\Unit\Plugins\Feedback\Support;
 
-use Plugins\User\API\DTOs\ListFeedbackQuery;
-use Plugins\User\Application\Ports\FeedbackStore;
-use Plugins\User\Domain\Entities\FeedbackEntry;
+use Plugins\Feedback\API\DTOs\ListFeedbackQuery;
+use Plugins\Feedback\Application\Ports\FeedbackStore;
+use Plugins\Feedback\Domain\Entities\FeedbackEntry;
 
 /**
  * In-memory FeedbackStore for service tests. Insertion order is newest-last;
